Revert "Feature collect plugin"
