+++ apps/api/modules/bkdata_databus.py
@@ -21,6 +21,7 @@ the project delivered to anyone in the future.
 """
 from django.utils.translation import ugettext_lazy as _
 
+
 from apps.api.base import DataAPI
 from apps.api.modules.utils import add_esb_info_before_request_for_bkdata_user
 from config.domains import DATABUS_APIGATEWAY_ROOT
@@ -34,26 +35,7 @@ class _BkDataDatabusApi:
             method="GET",
             url=DATABUS_APIGATEWAY_ROOT + "data_storages/",
             module=self.MODULE,
-            description="获取数据入库列表",
-            default_return_value=None,
-            before_request=add_esb_info_before_request_for_bkdata_user,
-            cache_time=60,
-        )
-        self.databus_data_storages_post = DataAPI(
-            method="POST",
-            url=DATABUS_APIGATEWAY_ROOT + "data_storages/",
-            module=self.MODULE,
-            description="创建入库",
-            default_return_value=None,
-            before_request=add_esb_info_before_request_for_bkdata_user,
-            cache_time=60,
-        )
-        self.databus_data_storages_put = DataAPI(
-            method="PUT",
-            url=DATABUS_APIGATEWAY_ROOT + "data_storages/{result_table_id}/",
-            module=self.MODULE,
-            description="更新入库",
-            url_keys=["result_table_id"],
+            description=u"获取数据入库列表",
             default_return_value=None,
             before_request=add_esb_info_before_request_for_bkdata_user,
             cache_time=60,
@@ -63,7 +45,7 @@ class _BkDataDatabusApi:
             method="GET",
             url=DATABUS_APIGATEWAY_ROOT + "cleans/",
             module=self.MODULE,
-            description="获取清洗配置列表",
+            description=u"获取清洗配置列表",
             default_return_value=None,
             before_request=add_esb_info_before_request_for_bkdata_user,
         )
@@ -71,7 +53,7 @@ class _BkDataDatabusApi:
             method="POST",
             url=DATABUS_APIGATEWAY_ROOT + "cleans/",
             module=self.MODULE,
-            description="创建清洗配置",
+            description=u"创建清洗配置",
             default_return_value=None,
             before_request=add_esb_info_before_request_for_bkdata_user,
         )
@@ -80,7 +62,7 @@ class _BkDataDatabusApi:
             url=DATABUS_APIGATEWAY_ROOT + "cleans/{processing_id}/",
             module=self.MODULE,
             url_keys=["processing_id"],
-            description="更新清洗配置",
+            description=u"更新清洗配置",
             default_return_value=None,
             before_request=add_esb_info_before_request_for_bkdata_user,
         )
@@ -88,7 +70,7 @@ class _BkDataDatabusApi:
             method="POST",
             url=DATABUS_APIGATEWAY_ROOT + "tasks/",
             module=self.MODULE,
-            description="创建清洗分发任务",
+            description=u"创建清洗分发任务",
             default_return_value=None,
             before_request=add_esb_info_before_request_for_bkdata_user,
         )
@@ -97,7 +79,7 @@ class _BkDataDatabusApi:
             url=DATABUS_APIGATEWAY_ROOT + "tasks/{result_table_id}/",
             module=self.MODULE,
             url_keys=["result_table_id"],
-            description="停止清洗，分发任务",
+            description=u"停止清洗，分发任务",
             default_return_value=None,
             before_request=add_esb_info_before_request_for_bkdata_user,
         )
