+++ apps/exceptions.py
@@ -36,7 +36,6 @@ class ErrorCode(object):
     BKLOG_EXTRACT = "26"
     BKLOG_COLLECTOR_CONFIG = "31"
     BKLOG_ESQUERY = "32"
-    BKLOG_COLLECTOR_PLUGIN = "33"
     BKLOG_TENCENT_AUTH = "41"
     BKLOG_UPGRADE = "51"
     BKLOG_MEASURE = "61"
