+++ apps/log_databus/constants.py
@@ -20,12 +20,13 @@ We undertake not to change the open source license (MIT license) applicable to t
 the project delivered to anyone in the future.
 """
 import markdown
+
 from django.conf import settings
 from django.utils import translation
 from django.utils.translation import ugettext_lazy as _
 
-from apps.feature_toggle.handlers.toggle import FeatureToggleObject
 from apps.utils import ChoicesEnum
+from apps.feature_toggle.handlers.toggle import FeatureToggleObject
 
 META_PARAMS_DATE_FORMAT = "%Y-%m-%d %H:%M:%S"
 RESTORE_INDEX_SET_PREFIX = "restore_"
@@ -308,17 +309,6 @@ class EtlConfig(object):
     BK_LOG_JSON = "bk_log_json"
     BK_LOG_DELIMITER = "bk_log_delimiter"
     BK_LOG_REGEXP = "bk_log_regexp"
-    CUSTOM = "custom"
-
-
-class EtlConfigChoices(ChoicesEnum):
-    _choices_labels = (
-        (EtlConfig.BK_LOG_TEXT, _("直接入库")),
-        (EtlConfig.BK_LOG_JSON, _("Json")),
-        (EtlConfig.BK_LOG_DELIMITER, _("分隔符")),
-        (EtlConfig.BK_LOG_REGEXP, _("正则")),
-        (EtlConfig.CUSTOM, _("自定义")),
-    )
 
 
 # 节点属性字段过滤黑名单
@@ -338,17 +328,3 @@ BKDATA_ES_TYPE_MAP = {
 }
 
 ETL_PARAMS = {"retain_original_text": True, "separator_regexp": "", "separator": ""}
-
-
-class ETLProcessorChoices(ChoicesEnum):
-    """
-    数据处理器
-    """
-
-    TRANSFER = "transfer"
-    BKBASE = "bkbase"
-
-    _choices_labels = (
-        (TRANSFER, _("Transfer")),
-        (BKBASE, _("数据平台")),
-    )
