+++ apps/log_databus/exceptions.py
@@ -34,11 +34,6 @@ class BaseCollectorConfigException(BaseException):
     MESSAGE = _("采集配置模块异常")
 
 
-class BaseCollectorPluginException(BaseException):
-    MODULE_CODE = ErrorCode.BKLOG_COLLECTOR_PLUGIN
-    MESSAGE = _("采集插件模块异常")
-
-
 class CollectorConfigNotExistException(BaseCollectorConfigException):
     ERROR_CODE = "001"
     MESSAGE = _("采集配置不存在")
@@ -47,11 +42,6 @@ class CollectorConfigNotExistException(BaseCollectorConfigException):
 class DataLinkConfigNotExistException(BaseCollectorConfigException):
     ERROR_CODE = "002"
     MESSAGE = _("链路配置不存在")
-
-
-class CollectorPluginNotExistException(BaseCollectorPluginException):
-    ERROR_CODE = "003"
-    MESSAGE = _("采集插件不存在")
 
 
 class CollectorIdNotExistException(BaseCollectorConfigException):
@@ -124,16 +114,6 @@ class CollectorResultTableIDDuplicateException(BaseCollectorConfigException):
     MESSAGE = _("采集项{result_table_id}结果表ID重复")
 
 
-class CollectorPluginNameDuplicateException(BaseCollectorPluginException):
-    ERROR_CODE = "115"
-    MESSAGE = _("采集插件名称已存在")
-
-
-class CollectorPluginNotMatchException(BaseCollectorPluginException):
-    ERROR_CODE = "116"
-    MESSAGE = _("参数异常：采集插件不匹配")
-
-
 class StorageNotExistException(BaseCollectorConfigException):
     ERROR_CODE = "201"
     MESSAGE = _("集群不存在")
@@ -177,11 +157,6 @@ class HotColdCheckException(BaseCollectorConfigException):
 class StorageHaveResource(BaseCollectorConfigException):
     ERROR_CODE = "209"
     MESSAGE = _("集群还有未删除的采集项、第三方集群索引集")
-
-
-class BKBASEStorageNotExistException(BaseCollectorPluginException):
-    ERROR_CODE = "210"
-    MESSAGE = _("集群未同步到数据平台")
 
 
 class EtlNotSupportedException(BaseCollectorConfigException):
