+++ apps/log_databus/handlers/etl.py
@@ -21,72 +21,51 @@ the project delivered to anyone in the future.
 """
 
 import arrow
-from django.conf import settings
-from django.db import transaction
-from django.utils.module_loading import import_string
-from django.utils.translation import ugettext_lazy as _
 
-from apps.api import TransferApi
+from django.utils.translation import ugettext_lazy as _
+from django.db import transaction
+from django.conf import settings
+
+from apps.constants import UserOperationTypeEnum, UserOperationActionEnum
 from apps.feature_toggle.handlers.toggle import FeatureToggleObject
 from apps.feature_toggle.plugins.constants import FEATURE_COLLECTOR_ITSM
-from apps.log_databus.constants import ETLProcessorChoices, ETL_PARAMS, EtlConfig, REGISTERED_SYSTEM_DEFAULT
+from apps.log_clustering.handlers.clustering_config import ClusteringConfigHandler
+from apps.log_clustering.handlers.data_access.data_access import DataAccessHandler
 from apps.log_databus.exceptions import (
     CollectorConfigNotExistException,
     EtlParseTimeFormatException,
     EtlStorageUsedException,
+    CollectorActiveException,
+    CollectorResultTableIDDuplicateException,
 )
+from apps.log_databus.handlers.collector_scenario import CollectorScenario
+from apps.log_databus.handlers.collector_scenario.custom_define import get_custom
 from apps.log_databus.handlers.etl_storage import EtlStorage
-from apps.log_databus.models import (
-    CollectorConfig,
-    ItsmEtlConfig,
-    StorageCapacity,
-    StorageUsed,
-)
-from apps.log_search.constants import FieldDateFormatEnum, ISO_8601_TIME_FORMAT_NAME
+from apps.log_databus.models import CollectorConfig, StorageCapacity, StorageUsed, CleanStash, ItsmEtlConfig
+from apps.log_clustering.tasks.flow import update_clustering_clean
 from apps.log_search.handlers.index_set import IndexSetHandler
-from apps.log_search.models import ProjectInfo, Scenario
+from apps.log_search.models import Scenario, ProjectInfo
+from apps.log_search.constants import FieldDateFormatEnum, CollectorScenarioEnum, ISO_8601_TIME_FORMAT_NAME
 from apps.models import model_to_dict
 from apps.utils.db import array_group
+from apps.log_databus.handlers.storage import StorageHandler
+from apps.log_databus.constants import REGISTERED_SYSTEM_DEFAULT, EtlConfig, ETL_PARAMS
+from apps.decorators import user_operation_record
+from apps.utils.local import get_request_username
+from apps.utils.log import logger
+from apps.api import TransferApi
 
 
 class EtlHandler(object):
-    def __init__(self, collector_config_id=None, etl_processor=ETLProcessorChoices.TRANSFER.value):
+    def __init__(self, collector_config_id=None):
         super().__init__()
         self.collector_config_id = collector_config_id
         self.data = None
-        self.etl_processor = etl_processor
         if collector_config_id:
-            self.data = self._get_collect_config(collector_config_id)
-            self.etl_processor = self.data.etl_processor
-
-    @staticmethod
-    def _get_collect_config(collector_config_id):
-        try:
-            collect_config: CollectorConfig = CollectorConfig.objects.get(collector_config_id=collector_config_id)
-            return collect_config
-        except CollectorConfig.DoesNotExist:
-            raise CollectorConfigNotExistException()
-
-    def __new__(cls, *args, **kwargs):
-        # 获取处理器信息
-        etl_processor = kwargs.get("etl_processor", ETLProcessorChoices.TRANSFER.value)
-        collector_config_id = kwargs.get("collector_config_id")
-        if collector_config_id:
-            collect_config = cls._get_collect_config(collector_config_id)
-            etl_processor = collect_config.etl_processor
-        # 处理器映射关系
-        mapping = {
-            ETLProcessorChoices.BKBASE.value: "BKBaseEtlHandler",
-            ETLProcessorChoices.TRANSFER.value: "TransferEtlHandler",
-        }
-        # 获取处理器
-        try:
-            etl_handler = import_string(
-                "apps.log_databus.handlers.etl.{}.{}".format(etl_processor, mapping.get(etl_processor))
-            )
-            return object.__new__(etl_handler)
-        except ImportError as error:
-            raise NotImplementedError(f"EtlHandler of {etl_processor} not implement, error: {error}")
+            try:
+                self.data = CollectorConfig.objects.get(collector_config_id=self.collector_config_id)
+            except CollectorConfig.DoesNotExist:
+                raise CollectorConfigNotExistException()
 
     def check_es_storage_capacity(self, cluster_info, storage_cluster_id):
         if self.data.table_id:
@@ -131,8 +110,104 @@ class EtlHandler(object):
             collect_config.set_itsm_success()
         return data, True
 
-    def update_or_create(self, *args, **kwargs):
-        raise NotImplementedError
+    def update_or_create(
+        self,
+        etl_config,
+        table_id,
+        storage_cluster_id,
+        retention,
+        allocation_min_days,
+        storage_replies,
+        view_roles=None,
+        etl_params=None,
+        fields=None,
+        username="",
+    ):
+        # 停止状态下不能编辑
+        if self.data and not self.data.is_active:
+            raise CollectorActiveException()
+
+        # 当清洗为直接入库时，直接清理对应采集项清洗配置stash
+        if etl_config == EtlConfig.BK_LOG_TEXT:
+            CleanStash.objects.filter(collector_config_id=self.collector_config_id).delete()
+
+        # 存储集群信息
+        cluster_info = StorageHandler(storage_cluster_id).get_cluster_info_by_id()
+        self.check_es_storage_capacity(cluster_info, storage_cluster_id)
+        is_add = False if self.data.table_id else True
+
+        if self.data.is_clustering:
+            clustering_handler = ClusteringConfigHandler(collector_config_id=self.data.collector_config_id)
+            ClusteringConfigHandler.pre_check_fields(
+                fields=fields, etl_config=etl_config, clustering_fields=clustering_handler.data.clustering_fields
+            )
+            if clustering_handler.data.bkdata_etl_processing_id:
+                DataAccessHandler().create_or_update_bkdata_etl(self.data.collector_config_id, fields, etl_params)
+            etl_params["etl_flat"] = True
+            log_clustering_fields = CollectorScenario.log_clustering_fields(cluster_info["cluster_config"]["version"])
+            fields = CollectorScenario.fields_insert_field_index(source_fields=fields, dst_fields=log_clustering_fields)
+            update_clustering_clean.delay(index_set_id=clustering_handler.data.index_set_id)
+
+        # 判断是否已存在同result_table_id
+        if CollectorConfig(table_id=table_id).get_result_table_by_id():
+            logger.error(f"result_table_id {table_id} already exists")
+            raise CollectorResultTableIDDuplicateException(
+                CollectorResultTableIDDuplicateException.MESSAGE.format(result_table_id=table_id)
+            )
+
+        # 1. meta-创建/修改结果表
+        etl_storage = EtlStorage.get_instance(etl_config=etl_config)
+        etl_storage.update_or_create_result_table(
+            self.data,
+            table_id=table_id,
+            storage_cluster_id=storage_cluster_id,
+            retention=retention,
+            allocation_min_days=allocation_min_days,
+            storage_replies=storage_replies,
+            fields=fields,
+            etl_params=etl_params,
+            es_version=cluster_info["cluster_config"]["version"],
+            hot_warm_config=cluster_info["cluster_config"].get("custom_option", {}).get("hot_warm_config"),
+        )
+
+        if not view_roles:
+            view_roles = []
+
+        # 2. 创建索引集
+        index_set = self._update_or_create_index_set(etl_config, storage_cluster_id, view_roles, username=username)
+
+        # add user_operation_record
+        operation_record = {
+            "username": username or get_request_username(),
+            "biz_id": self.data.bk_biz_id,
+            "record_type": UserOperationTypeEnum.ETL,
+            "record_object_id": self.data.collector_config_id,
+            "action": UserOperationActionEnum.CREATE if is_add else UserOperationActionEnum.UPDATE,
+            "params": {
+                "etl_config": etl_config,
+                "table_id": table_id,
+                "storage_cluster_id": storage_cluster_id,
+                "retention": retention,
+                "allocation_min_days": allocation_min_days,
+                "view_roles": view_roles,
+                "etl_params": etl_params,
+                "fields": fields,
+            },
+        }
+        user_operation_record.delay(operation_record)
+        if self.data.collector_scenario_id == CollectorScenarioEnum.CUSTOM.value:
+            custom_config = get_custom(self.data.custom_type)
+            custom_config.after_etl_hook(self.data)
+
+        return {
+            "collector_config_id": self.data.collector_config_id,
+            "collector_config_name": self.data.collector_config_name,
+            "etl_config": etl_config,
+            "index_set_id": index_set["index_set_id"],
+            "scenario_id": index_set["scenario_id"],
+            "storage_cluster_id": storage_cluster_id,
+            "retention": retention,
+        }
 
     @staticmethod
     def etl_preview(etl_config, etl_params, data):
