+++ apps/log_databus/models.py
@@ -20,13 +20,12 @@ We undertake not to change the open source license (MIT license) applicable to t
 the project delivered to anyone in the future.
 """
 
-from apps.exceptions import ApiResultError
-from apps.log_databus.exceptions import ArchiveNotFound
 from apps.models import MultiStrSplitByCommaFieldText
+from apps.log_databus.exceptions import ArchiveNotFound
+from apps.exceptions import ApiResultError
+from apps.utils.log import logger
 from apps.utils.cache import cache_one_hour
 from apps.utils.function import map_if
-from apps.utils.local import get_request_username
-from apps.utils.log import logger
 from apps.utils.thread import MultiExecuteFunc
 
 """
@@ -36,7 +35,7 @@ databus
 3. 入库
 """
 
-from django.db import models, transaction  # noqa
+from django.db import models  # noqa
 from django.utils import timezone  # noqa
 from django.utils.functional import cached_property  # noqa
 from django.utils.translation import ugettext_lazy as _  # noqa
@@ -44,8 +43,6 @@ from django_jsonfield_backport.models import JSONField  # noqa
 
 from apps.api import CmsiApi, TransferApi  # noqa
 from apps.log_databus.constants import (  # noqa
-    ETLProcessorChoices,
-    EtlConfigChoices,
     TargetObjectTypeEnum,  # noqa
     TargetNodeTypeEnum,  # noqa
     CollectItsmStatus,  # noqa
@@ -80,14 +77,12 @@ class CollectorConfig(SoftDeleteModel):
 
     collector_config_id = models.AutoField(_("采集配置ID"), primary_key=True)
     collector_config_name = models.CharField(_("采集配置名称"), max_length=64)
-    collector_plugin_id = models.BigIntegerField(_("采集插件ID"), db_index=True, null=True)
     bk_app_code = models.CharField(_("接入的来源APP"), max_length=64, default="bk_log_search")
     collector_scenario_id = models.CharField(_("采集场景"), max_length=64)
     custom_type = models.CharField(
         _("自定义类型"), max_length=30, choices=CustomTypeEnum.get_choices(), default=CustomTypeEnum.LOG.value
     )
     bk_biz_id = models.IntegerField(_("业务id"))
-    bkdata_biz_id = models.IntegerField(_("数据归属业务ID"), null=True)
     category_id = models.CharField(_("数据分类"), max_length=64)
     target_object_type = models.CharField(
         _("对象类型"), max_length=32, choices=TargetObjectTypeEnum.get_choices(), default=TargetObjectTypeEnum.HOST.value
@@ -103,14 +98,6 @@ class CollectorConfig(SoftDeleteModel):
     bk_data_id = models.IntegerField(_("采集链路data_id"), null=True, default=None)
     bk_data_name = models.CharField(_("采集链路data_name"), null=True, default=None, max_length=64)
     table_id = models.CharField(_("结果表ID"), max_length=255, null=True, default=None)
-    bkbase_table_id = models.CharField(_("BKBASE结果表ID"), max_length=255, null=True, default=None)
-    processing_id = models.CharField(_("计算平台清洗id"), max_length=255, null=True, blank=True)
-    etl_processor = models.CharField(
-        _("数据处理引擎"),
-        max_length=32,
-        choices=ETLProcessorChoices.get_choices(),
-        default=ETLProcessorChoices.TRANSFER.value,
-    )
     etl_config = models.CharField(_("清洗配置"), max_length=32, null=True, default=None)
     subscription_id = models.IntegerField(_("节点管理订阅ID"), null=True, default=None)
     task_id_list = MultiStrSplitByCommaField(_("最后一次部署任务"), max_length=255, null=True, default=None)
@@ -131,17 +118,6 @@ class CollectorConfig(SoftDeleteModel):
     storage_replies = models.IntegerField(_("ES副本数"), null=True, default=1, blank=True)
     bkdata_data_id_sync_times = models.IntegerField(_("调用数据平台创建data_id失败数"), default=0)
     collector_config_name_en = models.CharField(_("采集项英文名"), max_length=255, null=True, blank=True, default="")
-    is_display = models.BooleanField(_("采集项是否对用户可见"), default=True)
-
-    def get_bk_biz_id(self):
-        bk_biz_id = self.bkdata_biz_id if self.bkdata_biz_id else self.bk_biz_id
-        return bk_biz_id
-
-    def get_name(self):
-        return self.collector_config_name
-
-    def get_en_name(self):
-        return self.collector_config_name_en
 
     @property
     def is_clustering(self) -> bool:
@@ -368,7 +344,6 @@ class BKDataClean(SoftDeleteModel):
     bk_biz_id = models.IntegerField(_("业务id"))
     log_index_set_id = models.IntegerField(_("索引集id"), blank=True, null=True, db_index=True)
     is_authorized = models.BooleanField(_("索引集是否被授权"), default=False)
-    etl_config = models.CharField(_("清洗配置"), max_length=32, null=True, blank=True)
 
     class Meta:
         verbose_name = _("高级清洗列表")
@@ -487,90 +462,3 @@ class RestoreConfig(SoftDeleteModel):
     def get_collector_config_id(cls, restore_config_id):
         restore: "RestoreConfig" = cls.objects.get(restore_config_id=restore_config_id)
         return restore.archive.collector_config.collector_config_id
-
-
-class CollectorPlugin(SoftDeleteModel):
-    """
-    采集插件，控制采集项行为
-    """
-
-    bk_biz_id = models.BigIntegerField(_("业务ID"))
-    bkdata_biz_id = models.BigIntegerField(_("数据归属业务ID"), null=True, blank=True)
-    collector_plugin_id = models.BigAutoField(_("采集插件ID"), primary_key=True)
-    collector_plugin_name = models.CharField(_("采集插件名称"), max_length=64)
-    collector_plugin_name_en = models.CharField(_("英文采集插件名称"), max_length=64)
-    collector_scenario_id = models.CharField(_("采集场景ID"), max_length=64)
-    description = models.CharField(_("插件描述"), max_length=64)
-    category_id = models.CharField(_("数据分类"), max_length=64)
-    data_encoding = models.CharField(_("日志字符集"), max_length=30, null=True, default=None)
-    is_display_collector = models.BooleanField(_("采集项是否对用户可见"), default=False)
-    is_allow_alone_data_id = models.BooleanField(_("是否允许使用独立DATAID"), default=True)
-    bk_data_id = models.IntegerField(_("DATAID"), null=True)
-    data_link_id = models.IntegerField(_("数据链路ID"), null=True)
-    processing_id = models.CharField(_("计算平台清洗id"), max_length=255, null=True, blank=True)
-    is_allow_alone_etl_config = models.BooleanField(_("是否允许独立配置清洗规则"), default=True)
-    etl_processor = models.CharField(
-        _("数据处理器"), max_length=32, choices=ETLProcessorChoices.get_choices(), default=ETLProcessorChoices.TRANSFER.value
-    )
-    etl_config = models.CharField(
-        _("清洗配置"), max_length=32, null=True, default=None, choices=EtlConfigChoices.get_choices()
-    )
-    etl_params = models.JSONField(_("清洗参数"), null=True)
-    fields = models.JSONField(_("清洗字段"), null=True)
-    params = models.JSONField(_("采集插件参数"), default=dict, null=True)
-    table_id = models.CharField(_("结果表ID"), max_length=255, null=True)
-    bkbase_table_id = models.CharField(_("BKBASE结果表ID"), max_length=255, null=True)
-    is_allow_alone_storage = models.BooleanField(_("是否允许独立存储"), default=True)
-    storage_cluster_id = models.IntegerField(_("存储集群ID"), null=True)
-    retention = models.IntegerField(_("数据有效时间"), null=True)
-    allocation_min_days = models.IntegerField(_("冷热数据生效时间"), null=True)
-    storage_replies = models.IntegerField(_("副本数量"), null=True)
-    storage_shards_nums = models.IntegerField(_("ES分片数量"), null=True, default=None, blank=True)
-    storage_shards_size = models.IntegerField(_("单shards分片大小"), null=True, default=None, blank=True)
-
-    class Meta:
-        verbose_name = _("用户采集插件")
-        verbose_name_plural = verbose_name
-        ordering = ("-updated_at",)
-        unique_together = [
-            ("collector_plugin_name", "bk_biz_id"),
-            ("collector_plugin_name_en", "bk_biz_id"),
-        ]
-
-    def get_updated_by(self):
-        if self.updated_by == ADMIN_REQUEST_USER:
-            return self.created_by
-        return self.updated_by
-
-    def get_bk_biz_id(self):
-        bk_biz_id = self.bkdata_biz_id if self.bkdata_biz_id else self.bk_biz_id
-        return bk_biz_id
-
-    def get_name(self):
-        return self.collector_plugin_name
-
-    def get_en_name(self):
-        return self.collector_plugin_name_en
-
-    def get_transfer_table_id(self):
-        return self.transfer_table_id
-
-    def set_transfer_table_id(self, table_id: str):
-        self.transfer_table_id = table_id
-        self.save()
-
-    def get_table_id(self):
-        if self.etl_processor == ETLProcessorChoices.BKBASE.value:
-            return self.bkbase_table_id
-        return self.table_id
-
-    @transaction.atomic()
-    def change_collector_display_status(self, display_status: bool):
-        """更改采集项可见状态"""
-        request_user = get_request_username()
-        self.is_enabled_display_collector = display_status
-        self.save()
-        CollectorConfig.objects.filter(collector_plugin_id=self.collector_plugin_id).update(
-            is_display=display_status, updated_at=timezone.now(), updated_by=request_user
-        )
-        logger.info("[Change Collector Plugin Display Status] %s %s", self.collector_plugin_id, request_user)
