+++ apps/log_databus/urls.py
@@ -19,19 +19,13 @@ SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 We undertake not to change the open source license (MIT license) applicable to the current version of
 the project delivered to anyone in the future.
 """
-from django.conf.urls import include, url
+from django.conf.urls import url, include
 from rest_framework import routers
 
-from apps.log_databus.views import (
-    archive_views,
-    clean_views,
-    collector_plugin_views,
-    collector_views,
-    itsm_views,
-    restore_views,
-)
-from apps.log_databus.views import link_views
+from apps.log_databus.views import collector_views, itsm_views, archive_views, restore_views, clean_views
 from apps.log_databus.views import storage_views
+from apps.log_databus.views import link_views
+
 
 router = routers.DefaultRouter(trailing_slash=True)
 
@@ -39,7 +33,6 @@ router.register(r"archive", archive_views.ArchiveViewSet, basename="archive")
 router.register(r"restore", restore_views.RestoreViewSet, basename="restore")
 router.register(r"storage", storage_views.StorageViewSet, basename="databus_storage")
 router.register(r"collectors", collector_views.CollectorViewSet, basename="collectors")
-router.register(r"collector_plugins", collector_plugin_views.CollectorPluginViewSet, basename="collector_plugins")
 router.register(r"data_link", link_views.DataLinkViewSet, basename="data_link")
 router.register(r"collect_itsm", itsm_views.ItsmViewSet, basename="collect_itsm")
 router.register(r"collect_itsm_cb", itsm_views.ItsmCallbackViewSet, basename="collect_itsm_cb")
